Some big changes.
You now have to add some text on the TOP, of each script with the objects that are inside. You only do this for objects you want to show in the menu.
You nolonger need the class name for the setup of a object.

Objects.list holds a list of objects to show in the menu.

Now all the plugin compiler is going to do is compile.